don't string-smash errors if you don't have too
